share form instances across a context
