+++ apps/web/src/pages/completed.tsx
@@ -1,10 +1,10 @@
 import { FormList } from 'apps/web/src/components/FormList';
-import { useForm } from '@web/hooks/useForm';
 import { FormListPageLayout } from '@web/components/FormListPageLayout';
 import isAuth from '@web/components/isAuth';
+import { useUserFormsContext } from '@web/context/UserFormsContext';
 
 function Completed() {
-  const { completedForms } = useForm();
+  const { completedForms } = useUserFormsContext();
 
   return (
     <>
