+++ apps/web/src/pages/index.tsx
@@ -1,11 +1,11 @@
 import { Box, Text } from '@chakra-ui/react';
-import { useForm } from '@web/hooks/useForm';
 import Error from './../components/Error';
 import FormLoading from './../components/FormLoading';
 import { FormList } from './../components/FormList';
 import { useAuth } from '@web/hooks/useAuth';
 import { OverviewRow } from '@web/components/OverviewRow';
 import isAuth from '@web/components/isAuth';
+import { useUserFormsContext } from '@web/context/UserFormsContext';
 
 function Overview() {
   const {
@@ -15,7 +15,7 @@ function Overview() {
     assignedFIError,
     createdFILoading,
     createdFIError,
-  } = useForm();
+  } = useUserFormsContext();
 
   const { user } = useAuth();
